fix(mesh): round 4 Copilot PR review
1. **obj parser tab tolerance** — `mtllib` / `usemtl` detection used
   `startsWith("mtllib ")` (single-space prefix), which fails for valid
   OBJ files that separate the keyword from its argument with tabs or
   multiple spaces. Switched to the same `split(/\s+/)` tokenization
   the `v/vt/f` paths use, then check `parts[0]` against the keyword.

2. **Renderer ↔ CanvasRenderTarget circular import resolved** —
   `CanvasRenderTarget` imported `Renderer` for `Renderer.createCanvas`,
   while `Renderer` already imported `CanvasRenderTarget` (used in its
   constructor). The cycle works today via ES module hoisting but is
   brittle. Extracted the canvas allocator into a new
   `video/canvas_factory.js` helper that both can import without
   referencing each other. `Renderer.createCanvas` becomes a thin
   re-export of the helper, so the public-facing API is unchanged.
   `CanvasRenderTarget` and `TMXLayer` migrated to importing the
   helper directly.

3. **"Single draw call" wording corrected in 5 places** — Mesh.draw
   doc, vertexColors-field doc, mesh_batcher.addMesh doc,
   webgl_renderer.drawMesh doc, and CHANGELOG line 10 all overstated
   the guarantee. `MeshBatcher.addMesh` still chunks very large meshes
   across multiple flushes to respect vertex/index buffer limits —
   what multi-material rendering actually guarantees is **no extra
   draw calls per material** vs single-material rendering, not a
   blanket "single draw call regardless of size". Rephrased
   consistently across all five sites.

README bullet (root + packages/melonjs/) updated to mention
multi-material support in the existing 3D mesh feature line.
